Started a new Class-Library "`SimScanner`"
This library will allow you to collect airports and parking locations, as well as aircraft data. Started with implementing a BglReader to test parsing BGL files. First version can read a BGL file and list sections, and when it is an airport it will read the subsection and display ICAO and Region codes. Will go as far as collecting Gates and parkings with LatLongAlt, but first I need to scan add-on config files to gather all scenery files. For MSFS that will work a bit differently, but I already have the packages directory so don't expect too much trouble there.