Major cleanup in the `NumberTextBox` hierarchy
Been able to remove most of the field customizations in the MainWindow code. Clicking in the field selects the field contents, so when you start typing it will replace the current value, which is a lot easier while flying; no need to drag or double-click. You can set event names for when "Enter" is pressed and wheel-up or -down.